Put getArtist/getTitle in the correct helper file
Both are about beatmapset, not beatmap.
