+++ resources/js/beatmap-discussions/header.tsx
@@ -23,8 +23,8 @@ import { kebabCase, size, snakeCase } from 'lodash';
 import { deletedUser } from 'models/user';
 import core from 'osu-core-singleton';
 import * as React from 'react';
-import { getArtist, getTitle } from 'utils/beatmap-helper';
 import { makeUrl } from 'utils/beatmapset-discussion-helper';
+import { getArtist, getTitle } from 'utils/beatmapset-helper';
 import { classWithModifiers } from 'utils/css';
 import { trans } from 'utils/lang';
 import BeatmapList from './beatmap-list';
