+++ resources/js/beatmapset-panel/index.tsx
@@ -18,8 +18,8 @@ import { observer } from 'mobx-react';
 import core from 'osu-core-singleton';
 import * as React from 'react';
 import { Transition } from 'react-transition-group';
-import { getArtist, getDiffColour, getTitle, group as groupBeatmaps } from 'utils/beatmap-helper';
-import { showVisual, toggleFavourite } from 'utils/beatmapset-helper';
+import { getDiffColour, group as groupBeatmaps } from 'utils/beatmap-helper';
+import { getArtist, getTitle, showVisual, toggleFavourite } from 'utils/beatmapset-helper';
 import { classWithModifiers } from 'utils/css';
 import { formatNumber, formatNumberSuffixed } from 'utils/html';
 import { trans } from 'utils/lang';
