+++ resources/js/beatmapsets-show/header.tsx
@@ -14,8 +14,7 @@ import { disposeOnUnmount, observer } from 'mobx-react';
 import core from 'osu-core-singleton';
 import * as React from 'react';
 import { renderToStaticMarkup } from 'react-dom/server';
-import { getArtist, getTitle } from 'utils/beatmap-helper';
-import { downloadLimited, toggleFavourite } from 'utils/beatmapset-helper';
+import { downloadLimited, getArtist, getTitle, toggleFavourite } from 'utils/beatmapset-helper';
 import { classWithModifiers } from 'utils/css';
 import { formatNumber } from 'utils/html';
 import { trans } from 'utils/lang';
