+++ resources/js/profile-page/beatmap-playcount.tsx
@@ -7,7 +7,7 @@ import UserLink from 'components/user-link';
 import BeatmapPlaycountJson from 'interfaces/beatmap-playcount-json';
 import GameMode from 'interfaces/game-mode';
 import * as React from 'react';
-import { getArtist, getTitle } from 'utils/beatmap-helper';
+import { getArtist, getTitle } from 'utils/beatmapset-helper';
 import { formatNumber } from 'utils/html';
 import { trans } from 'utils/lang';
 import { beatmapUrl } from 'utils/url';
