+++ resources/js/profile-page/play-detail.tsx
@@ -8,7 +8,8 @@ import { SoloScoreJsonForUser } from 'interfaces/solo-score-json';
 import UserJson from 'interfaces/user-json';
 import * as React from 'react';
 import PpValue from 'scores/pp-value';
-import { getArtist, getTitle, rulesetName, shouldShowPp } from 'utils/beatmap-helper';
+import { rulesetName, shouldShowPp } from 'utils/beatmap-helper';
+import { getArtist, getTitle } from 'utils/beatmapset-helper';
 import { classWithModifiers } from 'utils/css';
 import { formatNumber } from 'utils/html';
 import { trans } from 'utils/lang';
