+++ resources/js/quick-search/beatmapset.tsx
@@ -5,7 +5,7 @@ import BeatmapsetCover from 'components/beatmapset-cover';
 import BeatmapsetJson from 'interfaces/beatmapset-json';
 import { route } from 'laroute';
 import * as React from 'react';
-import { getArtist, getTitle } from 'utils/beatmap-helper';
+import { getArtist, getTitle } from 'utils/beatmapset-helper';
 import { classWithModifiers } from 'utils/css';
 import { trans } from 'utils/lang';
